# Improve performance of the store.