--- a/.github/workflows/go-ci.yml
+++ b/.github/workflows/go-ci.yml
@@ -48,6 +48,6 @@ jobs:
       - name: Check out code into the Go module directory
         uses: actions/checkout@v2
       - name: Install golangci-lint
-        run: curl -sSfL https://raw.githubusercontent.com/golangci/golangci-lint/master/install.sh | sh -s -- -b $(go env GOPATH)/bin v1.24.0
+        run: curl -sSfL https://raw.githubusercontent.com/golangci/golangci-lint/master/install.sh | sh -s -- -b $(go env GOPATH)/bin v1.31.0
       - name: Lint
         run: $(go env GOPATH)/bin/golangci-lint run