--- a/docs/.vuepress/config.js
+++ b/docs/.vuepress/config.js
@@ -52,6 +52,7 @@ module.exports = {
           children: [
             '',
             'get-started',
+            'benchmarks',
           ]
         },
         {