--- a/go.mod
+++ b/go.mod
@@ -2,4 +2,8 @@ module github.com/sdslabs/kiwi
 
 go 1.14
 
-require github.com/wangjia184/sortedset v0.0.0-20200422044937-080872f546ba
+require (
+	github.com/stretchr/testify v1.6.1 // indirect
+	github.com/tidwall/buntdb v1.1.4
+	github.com/wangjia184/sortedset v0.0.0-20200422044937-080872f546ba
+)