# Summary of progress (commits accordingly)

## Cleanup locking-unlocking of mutex (4319279)

Most of the functions, including `Do`, locked mutex twice for the
store instead of just once. Changes have been benchmarked and here
are results on my (author of this commit) computer:

Previous benchmarks:

    BenchmarkUpdate-8        3082083          372 ns/op
    BenchmarkGet-8           3884070          307 ns/op

New benchmarks

    BenchmarkUpdate-8        3547335          335 ns/op
    BenchmarkGet-8           4487338          264 ns/op

To run benchmarks, use `go test -bench=. ./stdkiwi`. Benchmarking
is done for `"str"` type of value.

## Remove `defer` from all the methods (6a4e790)

Defer is useful yet very expensive when it comes to performance.
On looking at benchmarks, it's quite logical to remove defer
statements and writing a little "dirty" code since the difference
is significant.

Old benchmarks:

    BenchmarkUpdate-8        3547335          335 ns/op
    BenchmarkGet-8           4487338          264 ns/op

New benchmarks:

    BenchmarkUpdate-8        3897735          307 ns/op
    BenchmarkGet-8           4881136          242 ns/op

## Cache do-map to avoid unnecessary allocations (22954f6)

Each time an action is executed, space is allocated for the do map.
When a key is added or updated, the map is cached for the value so
while executing the action no heap allocation takes place.

Old Benchmarks

    BenchmarkUpdate-8        3897735          307 ns/op         336 B/op        7 allocs/op
    BenchmarkGet-8           4881136          242 ns/op         288 B/op        4 allocs/op

New Benchmarks

    BenchmarkUpdate-8       10086518          115 ns/op          48 B/op        3 allocs/op
    BenchmarkGet-8          21886411          53.4 ns/op          0 B/op        0 allocs/op

---

## Final comparison with BuntDB

```
❯ go test -bench=. -test.benchmem ./benchmark
goos: darwin
goarch: amd64
pkg: github.com/sdslabs/kiwi/benchmark
BenchmarkBuntDB_Update-8        11777931                96.6 ns/op            48 B/op          1 allocs/op
BenchmarkBuntDB_View-8          23310963                47.1 ns/op            48 B/op          1 allocs/op
BenchmarkKiwi_Update-8          10356004               115 ns/op              48 B/op          3 allocs/op
BenchmarkKiwi_Get-8             21910110                53.2 ns/op             0 B/op          0 allocs/op
PASS
ok      github.com/sdslabs/kiwi/benchmark       6.216s
```

Resolves #14